requirements.txt
yt-dlp==2023.12.30
openai-whisper==20231117
torch==2.1.0
langchain-community==0.0.13
langchain==0.1.0
requests==2.31.0
ffmpeg-python==0.2.0
numpy>=1.21.0
scipy>=1.7.0
